Search token on change
