+++ packages/extension/src/components/pages/AddNewTrustline/StepConfirm/index.ts
@@ -1,0 +1,1 @@
+export * from './StepConfirm';
