+++ packages/extension/src/components/pages/AddNewTrustline/StepForm/SearchToken/index.ts
@@ -1,0 +1,1 @@
+export * from './SearchToken';
