+++ packages/extension/src/components/pages/AddNewTrustline/StepForm/index.ts
@@ -1,0 +1,1 @@
+export * from './StepForm';
