+++ packages/extension/src/components/pages/AddNewTrustline/StepWarning/index.ts
@@ -1,0 +1,1 @@
+export * from './StepWarning';
