+++ packages/extension/src/components/pages/AddNewTrustline/index.ts
@@ -1,4 +1,1 @@
 export * from './AddNewTrustline';
-export * from './StepForm';
-export * from './StepWarning';
-export * from './StepConfirm';
